requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.25
pymysql==1.1.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
alembic==1.13.1
openpyxl==3.1.5
apscheduler==3.10.4
